+++ src/components/requests/RecipeRequests.tsx
@@ -31,6 +31,7 @@ import { ReferenceContext } from "../../context/ReferenceContext";
 import RecipeFormControls from "../ui/controls/RecipeFormControls";
 import Dropdown from "../ui/common/Dropdown";
 import { idText } from "typescript";
+import RecipeForm from "../views/RecipeForm";
 
 /** Processes recipe data. Context data is passed through here on edit. Else template data.
  * RecipeRequests data (e.g recipe state) is mutable while context data(reference data) is not
@@ -40,7 +41,7 @@ import { idText } from "typescript";
  * MainContainer -> RecipeRequests -> [IngredientsGroup, InstructionsArea, NotesInput, TitleInput]
  */
 function RecipeRequests({
-  recipeActions,
+  stateActions,
   closeDialog,
   isOpen,
 }: RecipeRequestsProps) {
@@ -59,7 +60,8 @@ function RecipeRequests({
   const [isDisabled, setIsDisabled] = useState(true);
   const [isCopyAuthed, setIsCopyAuthed] = useState(false);
   const [isBookSelectOpen, setIsBookSelectOpen] = useState(false);
-  const [flag, setFlag] = useState(false);
+
+  const dialogPanelRef = useRef(null);
 
   /** replaces dialog with dropdown */
   function openBookDropdown() {
@@ -69,7 +71,6 @@ function RecipeRequests({
   /** Updates state with selected book ID - changes book in UI*/
   async function selectBookId(id: number, book: Book) {
     setIsCopyAuthed(PRIVILEGES.sharedInbox);
-    setFlag(true);
     setUserData((user) => {
       const userData = { ...user };
       userData.currentBookId = id;
@@ -104,24 +105,6 @@ function RecipeRequests({
     setRecipe(selectedRecipe);
   }, [selectedRecipe.id]);
 
-  /** Enables/disables UPDATE submit */
-  useEffect(() => {
-    if (requestAction === "edit") {
-      const name = compareNames(selectedRecipe.name, recipe.name);
-      const ingredients = compareIngredients(
-        selectedRecipe.ingredients,
-        recipe.ingredients,
-      );
-      const instructions = compareInstructions(
-        selectedRecipe.instructions,
-        recipe.instructions,
-      );
-      const notes = compareNotes(selectedRecipe.notes, recipe.notes);
-      const isAltered = name || ingredients || instructions || notes;
-      setIsDisabled(!isAltered);
-    }
-  }, [recipe]);
-
   const isLargeScreen = useMediaQuery({ query: "(min-width: 1024px)" }); // lg breakpoint in Tailwind
 
   /** Updates recipe state */
@@ -141,7 +124,7 @@ function RecipeRequests({
         currentBookId,
         userId,
       );
-      recipeActions.updateRecipes(res);
+      stateActions.updateRecipes(res);
       return "submitted";
     } catch (error: any) {
       const message = errorHandling("RecipeRequests - addRecipe", error);
@@ -160,7 +143,7 @@ function RecipeRequests({
         selectedRecipe.id,
         mutatedData,
       );
-      recipeActions.editRecipe(res);
+      stateActions.editRecipe(res);
     } catch (error: any) {
       const message = errorHandling("RecipeRequests - editRecipe", error);
       setError(message);
@@ -180,7 +163,7 @@ function RecipeRequests({
         recipeId,
         selectedRecipe.created_by_id,
       );
-      if (res.message) recipeActions.deleteRecipe();
+      if (res.message) stateActions.deleteRecipe();
     } catch (error: any) {
       const message = errorHandling("RecipeRequests - deleteRecipe", error);
       setError(message);
@@ -192,7 +175,7 @@ function RecipeRequests({
   async function removeSharedRecipe(bookId: number, recipeId: number) {
     try {
       const res = await API.deleteSharedRecipe(bookId, recipeId);
-      if (res.message) recipeActions.deleteRecipe();
+      if (res.message) stateActions.deleteRecipe();
     } catch (error) {
       const message = errorHandling(
         "RecipeRequests - removeSharedRecipe",
@@ -218,30 +201,35 @@ function RecipeRequests({
     }
   }
 
-  async function handleDelete() {
+  async function deleteRecipe() {
     await deleteRecipe(userId, currentBookId, selectedRecipe.id);
   }
 
-  async function handleRemove() {
+  async function removeRecipe() {
     await removeSharedRecipe(currentBookId, selectedRecipe.id);
   }
 
-  async function handleSubmit(e) {
+  /** handle recipe submit */
+  async function submitRecipe(e) {
     e.preventDefault();
     const res = await addRecipe();
     if (res) closeDialog();
   }
-
-  const dialogPanelRef = useRef(null);
 
   function handleCloseDialog() {
     closeDialog();
     setTimeout(() => {
       // prevents flash of recipe copy controls
       setIsBookSelectOpen(false);
-      // updateRecipes(recipe);
     }, 50);
   }
+
+  const recipeActions = {
+    submit: submitRecipe,
+    remove: removeRecipe,
+    delete: deleteRecipe,
+    edit: editRecipe,
+  };
 
   return (
     <Dialog open={isOpen} onClose={handleCloseDialog} className="relative z-10">
@@ -265,90 +253,7 @@ function RecipeRequests({
                 isActionCopy={true}
               />
             ) : (
-              <>
-                {error && <Alert alert={error} degree={"yellow"} />}{" "}
-                {/* This will be a popup instead */}
-                {/* <form onSubmit={handleSubmit}> */}
-                <div className={requestAction !== "copyRemove" ? "h-80" : ""}>
-                  {/* <div className="mx-auto flex h-12 w-12 items-center justify-center rounded-full bg-green-100">
-                <CheckIcon aria-hidden="true" className="h-6 w-6 text-green-600" />
-              </div> */}
-                  {/* <div className="mt-3 h-full border-2 border-yellow-300 text-center sm:mt-5"> */}
-                  {/* <DialogTitle as="h3" className="text-base font-semibold leading-6 text-gray-900">
-                  Payment successful
-                </DialogTitle> */}
-                  {/* <div className="mt-2">
-                  <p className="text-sm text-gray-500">
-                    Lorem ipsum dolor sit amet consectetur adipisicing elit. Consequatur amet labore.
-                  </p>
-                </div> */}
-                  {requestAction === "copyRemove" && (
-                    <p>
-                      NOTE: Once a recipe is copied to a recipe book, you will
-                      be the owner of that copy.{" "}
-                    </p>
-                  )}
-
-                  {requestAction !== "copyRemove" && (
-                    <section
-                      id="RecipeRequests-book"
-                      className="mx-auto h-full flex-col "
-                    >
-                      <section
-                        id="RecipeRequests-recipe"
-                        className="flex h-2/3"
-                      >
-                        <ReferenceContext.Provider
-                          value={{ dialogPanelRef: dialogPanelRef }}
-                        >
-                          <section
-                            id="RecipeRequests-title-ingredients"
-                            className="flex-1 h-full flex flex-col"
-                          >
-                            <div className="">
-                              <TitleInput handleUpdate={handleRecipeUpdate} />
-                            </div>
-
-                            <div
-                              id="RecipeRequests-ingredients"
-                              className="flex-1 overflow-hidden"
-                            >
-                              <IngredientsGroup
-                                handleRecipeUpdate={handleRecipeUpdate}
-                              />
-                            </div>
-                          </section>
-
-                          <section
-                            id="RecipeRequests-instructions"
-                            className="flex-col flex flex-1 ml-4 rounded-md"
-                          >
-                            <InstructionsRequests
-                              handleRecipeUpdate={handleRecipeUpdate}
-                            />
-                          </section>
-                        </ReferenceContext.Provider>
-                      </section>
-
-                      <section id="RecipeRequests-notes" className="">
-                        <NotesInput handleUpdate={handleRecipeUpdate} />
-                      </section>
-                    </section>
-                  )}
-                  {/* </div> */}
-                </div>
-                <div className="SubmitButton mt-5 sm:mt-6">
-                  <RecipeFormControls
-                    handleSubmit={handleSubmit}
-                    isDisabled={isDisabled}
-                    recipe={recipe}
-                    handleRemove={handleRemove}
-                    handleDelete={handleDelete}
-                    editRecipe={editRecipe}
-                    openDropdown={openBookDropdown}
-                  />
-                </div>
-              </>
+              <RecipeForm />
             )}
             {/* </form> */}
           </DialogPanel>
